requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.52.0
anyio==4.9.0
apify-client==1.10.0
apify-shared==1.4.1
asyncpg==0.30.0
attrs==25.3.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dateutils==0.6.12
distro==1.9.0
fastapi==0.103.1
filetype==1.2.0
frozenlist==1.6.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-anthropic==0.3.13
langchain-core==0.3.61
langchain-google-genai==2.1.4
langchain-text-splitters==0.3.8
langsmith==0.3.42
MarkupSafe==3.0.2
more-itertools==10.7.0
multidict==6.4.4
numpy==2.2.6
oauthlib==3.2.2
openai==1.82.0
orjson==3.10.18
packaging==24.2
passlib==1.7.4
pip==24.3.1
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.0.0
pydantic_core==2.33.2
pydub==0.25.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9.1
scipy==1.15.3
six==1.17.0
passlib==1.7.4
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.0
starlette==0.46.2
tavily-python==0.7.2
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing_extensions==4.13.2
typing-inspection==0.4.1
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.22.0
yarl==1.20.0
zstandard==0.23.0
email-validator==2.2.0
dnspython==2.7.0
boto3

# Dev dependencies
pytest==7.0.0